Use ENV['ACCELA_ACCESS_TOKEN'] if present.
The access token itself is not a full-token (with expiry date, etc), so
returns a sparse token struct with just the access_token method.

This should make it more obvious if questions arise like 'why doesn't the
token respond to #scope'.

The access token in ENV will always take precedence, so it should be
changed before it expires.  To revert to the standard Token object, just
unset the environment variable ACCELA_ACCESS_TOKEN.